Make map generic over lock type
This is necessary since there is no particular lock implementation that
"makes sense" on `no_std`. This patch changes all types to be generic
over any lock type that implements `lock_api::RawMutex`, and defaults to
`parking_lot::RawMutex` on `std`.

Note that this patch forks the struct definition for `HashMap` into one
for `std` (which has a default for `L`) and one for `no_std` (which does
not). This is fine since any code written _without_ the `std` feature
_will_ compile with the `std` feature enabled (`indexmap` does the same;
the reverse is not true). We _could_ introduce an intermediate private
struct to avoid repeating the definition, but it would require writing
`self.inner` all over the place, which would be sad. This seemed
marginally cleaner.

Also, this diff makes me want implied bounds a lot:
https://github.com/rust-lang/rust/issues/44491

This is essentially a rebase of adb58e03352f8dcef085faf97b33b4c7f34c8110
and f9ad7236ab133544cc32f54c5c0454f08ce519ca on top of the previous
commit + https://github.com/jonhoo/flurry/pull/47.