Prep for release of 0.1.0
This removes the patch that incorporates crossbeam-rs/crossbeam#458,
since that PR also bumps the version number, and 0.9 doesn't exist on
crates.io. This in turn means that the miri test has to be disabled for
this release in particular. I'll push a commit after release that
restores it and the crossbeam patch.

This commit also adds a changelog!