Relax bounds on many methods to be minimal
This patch removes a large number of unnecessary bounds. Beyond the
obvious ones (like iterators do not require `Hash` and constructors do
not need `BuildHasher`), the biggest change is that the thread-safety
bounds are now _only_ placed on methods that modify the map. The
reasoning here is simple enough: if the threads-safety bounds do not
hold for K/V, then the map must be empty, and if the map is empty, the
read operations are fine even if the K/V type is not thread-safe.