Make HashMap::put return more info
Now, `put` will let you know both what it ended up doing (insert,
replace, or nothing if no_replacement is set). This enables it to be
used in more settings. The return value also now carries information
about both the previous and new value for the key, which may be useful
in certain contexts.